+++ src/vmm/tests/integration_tests.rs
@@ -297,6 +297,7 @@ fn verify_load_snapshot(snapshot_file: TempFile, memory_file: TempFile) {
             network_overrides: vec![],
             vsock_override: None,
             clock_realtime: false,
+            drive_overrides: vec![],
         }))
         .unwrap();
 
@@ -383,6 +384,7 @@ fn verify_load_snap_disallowed_after_boot_resources(res: VmmAction, res_name: &s
         network_overrides: vec![],
         vsock_override: None,
         clock_realtime: false,
+        drive_overrides: vec![],
     });
     let err = preboot_api_controller.handle_preboot_request(req);
     assert!(